refactor: :recycle: Separated concerns better for state management by extracting it to its own file
