+++ src/main.tsx
@@ -2,32 +2,9 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { RouterProvider, createRouter } from "@tanstack/react-router";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { create } from "zustand";
-import { createJSONStorage, persist } from "zustand/middleware";
 import "./index.css";
-import MangaState from "./models/StoreState";
 import { routeTree } from "./routeTree.gen";
 import NotFoundComponent from "./shared/NotFoundComponents";
-
-export const useMangaStore = create<MangaState>()(
-  persist(
-    (set) => ({
-      favourites: [] as string[],
-      currentlyReading: ["", ""],
-      addToFavourites: (slug) =>
-        set((state) => ({ favourites: [...state.favourites, slug] })),
-      removeFromFavourites: (slug) =>
-        set((state) => ({
-          favourites: state.favourites.filter(
-            (favouriteHid) => slug != favouriteHid
-          ),
-        })),
-      setLastRead: (mangaSlug, mangaHid) =>
-        set(() => ({ currentlyReading: [mangaSlug, mangaHid] })),
-    }),
-    { name: "MangaState", storage: createJSONStorage(() => localStorage) }
-  )
-);
 
 const router = createRouter({
   routeTree,
