+++ src/modules/ReadMangaPage/ReadMangaPage.tsx
@@ -3,8 +3,8 @@ import { Link } from "@tanstack/react-router";
 import { useEffect } from "react";
 import { LazyLoadImage } from "react-lazy-load-image-component";
 import { getChapterByHid } from "../../api/api-services";
-import { useMangaStore } from "../../main";
 import { Route } from "../../routes/read.$manga.$chapter.lazy";
+import { useMangaStore } from "../../state/state-service";
 import checkImage from "../../utils/check-image-exists";
 import convertToUrl from "../../utils/convert-image-string";
 import BottomNavChaptersBar from "./components/BottomNavChaptersBar";
