+++ src/state/state-service.ts
@@ -1,0 +1,23 @@
+import { create } from "zustand";
+import { createJSONStorage, persist } from "zustand/middleware";
+import MangaState from "../models/StoreState";
+
+export const useMangaStore = create<MangaState>()(
+  persist(
+    (set) => ({
+      favourites: [] as string[],
+      currentlyReading: ["", ""],
+      addToFavourites: (slug) =>
+        set((state) => ({ favourites: [...state.favourites, slug] })),
+      removeFromFavourites: (slug) =>
+        set((state) => ({
+          favourites: state.favourites.filter(
+            (favouriteHid) => slug != favouriteHid
+          ),
+        })),
+      setLastRead: (mangaSlug, mangaHid) =>
+        set(() => ({ currentlyReading: [mangaSlug, mangaHid] })),
+    }),
+    { name: "MangaState", storage: createJSONStorage(() => localStorage) }
+  )
+);
